Further refactoring and create class that handle of maintaining optimization mappings
